requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
chromadb
langchain-google-genai